+++ apps/3000-home/next.config.js
@@ -11,7 +11,6 @@ const nextConfig = {
     // See: https://github.com/gregberge/svgr
     svgr: false,
   },
-
   webpack(config, options) {
     const { isServer } = options;