+++ apps/3000-home/remote-delegate.js
@@ -1,13 +1,27 @@
-import { importDelegatedModule } from '@module-federation/utilities';
+/* eslint-disable no-undef */
+const { importDelegatedModule } = require('@module-federation/utilities');
 
-module.exports = new Promise(async (resolve, reject) => {
+// Delegates are currently not used in this example, but are left here for testing.
+module.exports = new Promise((resolve, reject) => {
+  //eslint-disable-next-line
+  console.log('Delegate being called for', __resourceQuery);
+  //eslint-disable-next-line
   const currentRequest = new URLSearchParams(__resourceQuery).get('remote');
-  await new Promise((resolve) => setTimeout(resolve, 1000));
+
   const [global, url] = currentRequest.split('@');
 
-  const container = await importDelegatedModule({
+  importDelegatedModule({
     global,
     url: url + '?' + Date.now(),
   })
-    resolve(container);
+    .then(async (remote) => {
+      console.log(
+        __resourceQuery,
+        'resolved remote from',
+        __webpack_runtime_id__
+      );
+
+      resolve(remote);
+    })
+    .catch((err) => reject(err));
 });