+++ apps/3001-shop/next.config.js
@@ -12,7 +12,6 @@ const nextConfig = {
     // See: https://github.com/gregberge/svgr
     svgr: false,
   },
-  // swcMinify: true,
   webpack(config, options) {
     const { isServer } = options;