+++ packages/native-federation-tests/src/interfaces/HostOptions.ts
@@ -1,0 +1,6 @@
+export interface HostOptions {
+  moduleFederationConfig: any;
+  testsFolder?: string;
+  mocksFolder?: string;
+  deleteTestsFolder?: boolean;
+}